Collection Creation refacto ok, including middleware creation
